pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "dateparser~=1.2.0",
    "discord.py>=2.3.2",
    "google-generativeai>=0.5.0",
    "python-dotenv~=1.0.1",
    "pytz>=2024.1",
    "aiohttp~=3.9.5",
]